Add BaseMatcher for test specs
Summary:
This adds a `BaseMatcher` which will be used in the following diff as part of `TestSpec` generation.

It requires increasing visibility of certain internal methods. I tried to keep this to a minimum by having `BaseMatcher` itself sit in `com.facebook.litho`, allowing us to at least use package visibility where possible.

The matcher has a separate Builder. This split is done to avoid internal methods showing up in auto-complete when setting up a matcher in your test.

Right now, the matcher only has support for event handlers. In the future, this will be broadened to hopefully all common and other props, but as this is very repetitive, I will look into ways to generate that code either through annotation processors or through scripts.

Reviewed By: marco-cova

Differential Revision: D6834999

fbshipit-source-id: 669883e665abe6d8b287ff6b3b8eca7509b4e93a