Generate testspec matchers for event handlers
Summary:
This diff makes use of the previously added `BaseMatcher` infrastructure. The matchers in a testspec now extend `BaseMatcher` and run the common matchers upon construction. This can be seen in the generated specs as part of the integration test.

It requires some rearranging of deps as `BaseMatcher` is now required to be on the classpath when running tests.

Reviewed By: marco-cova

Differential Revision: D6844838

fbshipit-source-id: b67f469c9500db4bbef3bd91ac75603e019a85aa