Fix animating LithoView bounds on first mount
Summary:
When you try to animate `LithoView`'s height when it's first loaded in the Activity it's height remains 0 and animation will not run. In `LithoView.onMeasure()` we override default behavior when we see that height animation was specified and apply the current value (which is equal to 0 when Activity with LithoView is just appearing). And then animation is not triggered because ComponentTree is not mounted which is happening because incremental mount mounts an item when it has visible area which is not the case in our scenario.

Adding a condition to mount the component when there is LithoView bounds animation and it's height/width is 0 fixes the issue.

Reviewed By: astreet

Differential Revision: D6968169

fbshipit-source-id: 74b82204d724589f8855ac3a46b30964462af9fa