Experiment: Don't recalculate LayoutState synchronously in measure if only the root has changed (but size specs have stayed the same)
Summary:
Here's my rationale for why we can do this:

The root can change for two reasons: setRoot(Sync) and setRootAsync. If setRoot(Sync) was called, then we synchronously computed a layout state so it must match the main thread LayoutState. This means if the root component doesn't match in onMeasure, setRootAsync was called.

If setRootAsync was called, then we know we kicked off a BG layout with the size of the LithoView at the time. If the size changes, then we will synchronously computer a new main thread LayoutState, getting the new root automatically. If update state is called, we will similarly use the new root. Otherwise, we can just wait for the bg layout to complete to update the content, as we wanted to originally do.

This change is desireable because it lets us keep more things off the main thread.

Reviewed By: jsendros

Differential Revision: D7013159

fbshipit-source-id: ba1bfbe025d00716b0ce3df6664efca917d6a77f