Fix broken homepage code sample
Summary:
Well, this is embarrassing. This updates the frontpage example to actually
compile ...

I tried to make this reproducible and used src2png/prism and created a
stylesheet which more or less mirrors what we had before. It's not 1:1, but at
least it's pretty straightforward to create a new image now.

Reviewed By: muraziz

Differential Revision: D7023603

fbshipit-source-id: 0a5e7a68ac72c6aae653b23964001088ef458405