Enable clipbounds of TextSpec by default
Summary: To have a deterministic behavior of automatic bounds animation we need to clip drawables by their bounds. Recently we added an optional clipping of `TextDrawable` by bounds which is passed to drawable via `Text` component prop which was not clipped by default. This diff turns that default flag to `true`.

Reviewed By: marco-cova

Differential Revision: D6987841

fbshipit-source-id: d5e757d3f344df1d1bf5fcf6c0ae9ab7ac38108d