Remove customViewType duplicate check
This check is too aggressive for customViewTypes as it will always throw if a custom view type is set, since we require customViewType to bet set on all RenderInfos and they will be positively identified as duplicates with other RenderInfos of the same type.

For proper duplicate checking, we could check if there are Multiple ViewCreators bound to the same viewType, or if there are multiple viewTypes for the same ViewCreator. But in my case I would actually have different ViewCreator instances that may have the same behavior, and would have to implement a proper .equals() implementation. I think given the advanced API and the implied dangers of the API, we can simply leave these checks as-is without stricter checks.