Compare collections of components by using Component#isEquivalentTo
Summary:
When comparing props in component, we use `Component#isEquivalentTo` if it's a component object, and  `Object#equals` for others. However, if the prop is an instance of collection, we always use `Collection.equals` even if it contains components.

What we want to do in this diff is:
1. If the prop is a collection of components, iterate each element and use `isEquivalentTo` for comparison;
2. Otherwise, use `equals` method instead.

To achieve this, first we have to know if a object is an instance of collection and its type argument is an instance of component.
1. Because `java.util.Collection` is an interface, we add a field `mSuperinterfaces` in DeclaredTypeSpec.java to store type's super interfaces.
2. We also update the type arguments in DeclaredTypeSpec.java then we could know if it contains an component instance.

Then in `ComponentBodyGenerator.java` we can leverage `DeclaredTypeSpec#isSubInterface` and `DeclaredTypeSpec#getTypeArguments` to check if it's a collection of components and give it corresponding equal method.

Reviewed By: passy

Differential Revision: D7026181

fbshipit-source-id: ef0fb913684c2c79ac5b9c1278f96856d41afd87