Don't recreate ComponentTree in VerticalScrollSpec
Summary: creating a new ComponentTree every time in prepare was discarding any updated state for this. Also changed the measure logic to measure the component instead of the tree, because the ComponentTree wouldn't get remeasured if it already had measure specs, resulting in always receiving a sizeSpec with negative values from childComponentTree.setSizeSpec when we didn't recreate the tree.

Reviewed By: muraziz

Differential Revision: D7098988

fbshipit-source-id: dffe58b69c5c9dcc3741eb49004950a0cae7306d