Fix flaky LithoViewTestHelperTest
Summary:
The object identity hash code does not have a defined size. It's simply the
hexadecimal representation of the memory address which usually happens to be 7
or 8 digits long. This causes some flakiness in the test, with failures like
this:

```
Expecting:
  "litho.InlineLayout{94db65f V.E..... .. 0,0-100,100}
  litho.TestDrawableComponent{6f8d2d78 V.E..... .. 0,0-100,100 litho:id/test-drawable}
  litho.Text{5ca3ffa5 V.E..... .. 0,100-100,100 text="Hello, World"}"
to contain pattern:
  "litho.InlineLayout\{\w{8} V.E..... .. 0,0-100,100\}
  litho.TestDrawableComponent\{\w{8} V.E..... .. 0,0-100,100 litho:id/test-drawable\}
  litho.Text\{\w{8} V.E..... .. 0,100-100,100 text="Hello, World"\}"
```

This instead skips over any series of characters.

/ref D7084371

Reviewed By: mihaelao

Differential Revision: D7110155

fbshipit-source-id: d2570e212ab2ac06fc052d18fc6c57cb99bd33ff