Add ability to ask whether ComponentTree has completed the last requested layout in ComponentTreeHolder
Summary: Holds onto the last requested size specs and checks to see if the ComponentTree has a compatible layout for them when asked. This will be used when we get notified that a new layout has been calculated in order to understand whether that ComponentTree is ready to be added to the adapter.

Reviewed By: mihaelao

Differential Revision: D7070425

fbshipit-source-id: 9d59f148d16f1daf3cb6ef4ea31598bb4bbcfbc5