VerticalScroll: Measure component instead of component tree
Summary:
talked to Pasquale, we want to measure the componentTree instead of the component because when it needs to be mounted we already have measurements for it and it's more efficient.
This also fixes setting the LithoScrollView height to the height of the content. Basically went back to the implementation before D7098988 landed, and made sure we set a size spec in the output even when the ComponentTree doesn't need to be remeasured (otherwise the math.min check in measureVerticalScroll would always try to set a negative height on the tree if the tree had been previously measured).

Reviewed By: nisardan

Differential Revision: D7123747

fbshipit-source-id: 3fecb03e85c315d1397b21a020a5c7e13edaa428