Add generator for synthetic error event handler
Summary:
This adds a generator which creates a virtual event handler method specmodel
representation. This allows us to reuse the existig infrastructure for handling
events and generating the components code.

This will be wired up in diffs later in the stack but given the tests should be
easy to review like this.

For a quick idea of what the end result of this stack is supposed to be, take a look at the test in D7112854.

Reviewed By: IanChilds

Differential Revision: D7099597

fbshipit-source-id: a04f1cef05e0b7eb9db441c9688db8f3eb7ae7b1