Change onError delegate signature
Summary:
As outlined in T25182564, this is changing to a style that's closer to the React
model, relying on state to propagate changes. Because of that, we no longer
return a Component.

This temporarily breaks error handling altogether but as we're moving to a new model and it was always gated behind a feature flag, I consider this fair game. I'm also disabling some tests here which will be re-enabled later.

Reviewed By: IanChilds

Differential Revision: D7110244

fbshipit-source-id: a5ba632ca013bdcbe54c76577cd759aa5e430514