Generate error event handler
Summary:
This makes use of the previous diff and generates an `__internalOnErrorHandler`
method in components whose specs implement an `OnError` delegate.

This has the advantage of us reusing many pieces of existing infrastructure like event handlers and event dispatchers while maintaining a consistent look and feel for the high-level, user-facing API.

T26492923 tracks adding a validator to disallow usage of `__` in event handler names.

Reviewed By: marco-cova

Differential Revision: D7110326

fbshipit-source-id: 809a39c053da0ff2882141d011490423ae724b9b