Implement component-level error event dispatcher setup
Summary:
This sets up an event handler per component which dispatches to either the component's parent or, if it hits the root of the hierarchy, a default handler which reraises the exception captured within the event.

The event handler is stored within the component itself as opposed to `CommonProps` or `InternalNode` as it's neither a user-controlled prop nor in any way related to the underlying Android view infrastructure. It is closest to the `LoadingEvent` handler which exists in [Section](https://github.com/facebook/litho/blob/e7bc4e40aca7232ddddcf875c56706a5acfda254/litho-sections-core/src/main/java/com/facebook/litho/sections/Section.java#L44).

Reviewed By: marco-cova

Differential Revision: D7086135

fbshipit-source-id: d586c439c6df4daac2449cdd1a90bd452d6f7782