Integration test for error handling
Summary:
Setting up a test for onError in ComponentLifecycleTest is too tedious and would
not provide any meaningful signal as it would require mocking essentially the
entire world.

I'm setting this up as an integration test instead that tests the behavior end
to end instead by looking at the actual generated code.

Arguably, this should be part of the previous diff but as this touches a bunch of files, it's a lot easier to review this way and I will land the entire stack at once anyway.

Reviewed By: marco-cova

Differential Revision: D7112854

fbshipit-source-id: 6754ecaa3625c937c8b5788dfef758e1722cd33e