Get rid of mHasNonLazyUpdate
Summary:
The purpose of `mHasNonLazyUpdate` was to track whether we had any new non-lazy state updates when we finished applying a changeset, so that we could kick off a new update if we did, and just finish if we did not. However, it wasn't threadsafe, so it was possible to finish applying a new changeset and think that it had applied all the state updates, when in fact it had not, and so the latest change set would be ignored.

Instead, we keep track of all non-lazy state updates as well as all state updates, and we can use the map of non-lazy state updates to determine if all non-lazy state updates have been executed.

I can potentially move some more methods to StateUpdatesHolder; however, I thought it would be easier to review like this. Let me know if you'd rather I moved the methods there.

Reviewed By: mihaelao

Differential Revision: D7166920

fbshipit-source-id: 8c2b621f6d7a1781f0001e8355010a040a4688fd