Fix setSizeSpec with output
Summary:
One of those rare bug fixes that just involves deleting code ;)

Previously, we weren't setting output in setSizeSpec, and then in D7123747 we started setting incorrect output for this sizespec. Unfortunately, we do need to still calculate a layout in order to figure out the proper width and height of the component -- we can't just get it off the spec.

This diff removes some "protection" we had to make sure that setSizeSpec didn't conflict with the attached View's size, but this isn't useful in the majority of uses of this method where we're actually trying to measure content for a custom mountspec during a new layout calculation. In that case, either this measurement will match, or if it doesn't we'll get onBoundsDefined later anyway which will enforce the LithoView and content's measurements sync up again.

Reviewed By: pasqualeanatriello

Differential Revision: D7198201

fbshipit-source-id: ef4e04b9e33304c56b218edb50ac9d74a2cdae51