Refactor setting up background threads pool MC params
Summary:
bear with me while I change this :)
I'm trying to make it easier to configure these params for the experiment where we split the same layout, which will use similar params. There's now a LayoutThreadPoolConfiguration class that receives the params that exist in MC and sets up the pool size accordingly. We'll be using this to pass the pool size and thread priority to the LayoutThreadPoolExecutor

Reviewed By: astreet

Differential Revision: D7304380

fbshipit-source-id: 7bc400aa466efdd3bcad1ae8e4eab7833b4e152e