Provide WorkingRangeStatusHandler to record the working range status of component in a ComponentTree.
Summary:
WorkingRangeStatusHandler is used to store status of components with given working range. The key of the container is combined with the component's global key and the working range name.

The global key guarantees the uniqueness of the component in a ComponentTree, and it's consistent across different {link LayoutState}s. The working range name is used to find the specific working range since a component can have several working ranges. The value is an integer presenting the status.

 ---
Why global key is necessary?

Consider a case that ComponentA and ComponentB are both in a ComponentTree and have the same named range.
```
class ComponentASpec {
  OnCreateLayout
  static Component onCreateLayout() {
    ComponentA.registerPrefetchWorkingRange(c, new WorkingRange(...));
  }

  OnEnteredRange(name = "prefetch")
  static void rangeA() { ... }
}
```
```
class ComponentBSpec {
  OnCreateLayout
  static Component onCreateLayout() {
    ComponentA.registerPrefetchWorkingRange(c, new WorkingRange(...));
  }

  OnEnteredRange(name = "prefetch")
  static void rangeB() { ... }
}
```
Since ComponentA and ComponentB can define their own working range when registering a working range, we cannot guarantee a name in different components can map to the same working range. It means a range name in different components might have different working range status.

Reviewed By: mihaelao

Differential Revision: D7320985

fbshipit-source-id: 678d5f5e935974111a7f2a6d6cd2c87b85bcd8b1