Provide WorkingRangeContainer to record component working range
Summary:
WorkingRangeContainer is used to record all components annotated with `OnEnteredRange` and `OnExitedRange` in a component tree.
When a root component is entering/exiting the range, WorkingRangeContainer is used to notify descendants of the component.

WorkingRangeContainer a key-value mapping container that the key is composed of working range hashCode() and the member value of `name` in annotation.
For example:
```
LayoutSpec
OnCreateLayout
public class MyComponentSpec {
  OnCreateLayout
  static Component onCreateLayout(ComponentContext c, Prop MyWorkingRange myWorkingRange) {
    MyComponent.registerPrefetchWorkingRange(c, myWorkingRange);
    ...
  }

  OnEnteredRange(name = "prefetch")
  static void enterRangeFunction(ComponentContext c, ...) {
  }
}
```

The key is `"prefetch_" + myWorkingRange.hashCode()`, and the value is a wrapper that contains the working range and components information.

Reviewed By: mihaelao

Differential Revision: D7197786

fbshipit-source-id: 3a0d2975d6a2b44e2589e56cca7dfdfb7e87c58a