Refactor layout config so we can use logging
Summary: To use ComponentsLogger in the config module, we'd either need to put all the logging classes in a logging module or have this workaround. The first one is a bit more instrusive and it ends up in circular dependency that's not really worth getting around for now. This makes LayoutThreadPoolConfiguration an interface in the config module and moves the default implementation to the litho module.

Reviewed By: astreet

Differential Revision: D7427284

fbshipit-source-id: fb165360097a151d54e3c91b0792b57f6e2964c4