Generate accessors for injected fields
Summary:
This is the last step to get this rolling. It introduces a new interface method to the DI helper which generates an accessor method for obtaining the value for test purposes.

This can be necessary (and - surprise! - is in our case) when the DI container generates code which can't be generically accessed (e.g. not a package-private field, but a lazy access wrapper). This accessor method is written to the spec and can be stripped out by build tooling in prod environments. For tests, the method name is used from the test spec to obtain the value.

Reviewed By: IanChilds

Differential Revision: D7427515

fbshipit-source-id: a01c3af7c08618c26bd5f3e32869b39ecf0ac557