Pool views created from different ViewCreator instances in ViewCompatComponent separately
Summary:
Initially when  `ViewCompatComponent` was introduced `View`s created by `ViewCompatComponent` were pooled by `ViewCreator` instances which is the correct behaviour. After numerous changes into the framework itself including `ComponentsPools` mechanism and moving from `ComponentLayout` to `Component` that behaviour was no longer the case. Instead, the views were pooled by `ViewCreator` type, not an instance. That approach works as long as `ViewCreator`s provided to `ViewCompatComponent` had different type (like when you create an anonymous instance in which case the type (classname) will be different). But in many cases you create some class that implements `ViewCreator` and provide different instances of that class to `ViewCompatComponent` in which case all the views are recycled into the same pool which is where bug is coming from.

This diff fixes that behavior and aligns it to the original one. Now `ViewCompatComponent`s with different instances of `ViewCreator` will get different typeIds which will ensure that the `View`s they create are recycled into correct pools.

Reviewed By: IanChilds

Differential Revision: D7504579

fbshipit-source-id: f6b1e43c666adb4fec2cea98b818cd888a5860fd