Add extra check for creating a thread pool with invalid size
Summary:
Added a method that checks whether a device has multiple cores, which we'll use in the next diff to prevent creating a thread pool with invalid size.
This could have happened for single-core devices or for cases when we can't read the number of cores and the pool size multiplier was, for instance, 0.5. This resulted in a pool size of size 0, which is not allowed

since the number of cores doesn't change for a device, we can only calculate it once.

Reviewed By: astreet

Differential Revision: D7502004

fbshipit-source-id: 198206cfe9cc7da1a24b3acd5e2cd9ae0e1e4452