add support for OnUpdateStateWithTransition: Wire generated class/methods to Layout/Mount lifecycle
Summary:
The idea is simple when layout has been calculated and `LayoutState.collectResults` goes through all components to create `LayoutOutput`s we detect that component has `Transition` in its `StateContainer`. We save it into `TransitionContext` as pending state transition and when mount happens and tree is dirty we setup transitions and if there is pending state transition then we ignore the default one and apply that instead.

[UPDATE] After discussions, changed to this: We collect transitions from component's `StateContainer` when committing state changes into `ComponentTree`'s `StateHandler`. Later when mounting tree we will check whether we have pending transitions, and if so apply them when setting up transitions on dirty mount.

Reviewed By: astreet

Differential Revision: D7260359

fbshipit-source-id: 99803078c04ceb10d72a48dc04362d011e9f27a1