Prepare looper for thread in layout threadds pool
Summary:
If a new hander is being created inside a thread from the layout theads pool, it will crash with a RuntimeException: Can't create handler inside thread that has not called Looper.prepare(). This fixes that.
We already prepare the looper for the split layout executor, this only crashes when we use multiple background threads

Reviewed By: astreet

Differential Revision: D7585392

fbshipit-source-id: 201c233053c53a30a1d48820fc684c5b64be2585