Resubmission of 'Generate accessors for injected fields'
Summary:
Backout of original commit changeset: 041397611a16.

This didn't apply cleanly due to the introduction of isLazy on the InjectPropModel which in itself is a bit unfortunate as it is very implementation-specific. I mirror this now in my extractor spec.

The original commit was reverted due to a size regression. It turned out that the `VisibleForTesting` annotation is used as alias for `DoNotStrip` in certain redex configurations. D7568209 shows the rebased diff without removing the `VisibleForTesting` annotation.

Reviewed By: marco-cova

Differential Revision: D7585044

fbshipit-source-id: ceeb26ee1bd66b392dd344373645908b9fb6866d