Only perform fillViewport in the background when we need to (re)init the range
Summary:
First of all, there's no advantage to ever filling the viewport on the main thread.

Second, we want to fill the viewport only when we need to init a new range, meaning we're either being measured for the first time with content, or being remeasured (with content).

This also will depend on Jing-Wei's D7599886 that puts remeasures in the background. We will run our experiments together.

Reviewed By: mihaelao

Differential Revision: D7616487

fbshipit-source-id: cf57862e5e594701595a953422f8cb0bd36b8328