Adding dumping of ComponentTree from ComponentContext helper class
Summary:
This adds the functionality for dumping to a string the hierarchy of a given ComponentTree related to the provided ComponentContext.
This would help the debugging efforts across teams working in FB4A.

The functionality is going to be gated to debug builds due to current limitations of the framework.

Reviewed By: passy, astreet

Differential Revision: D7595856

fbshipit-source-id: dc553e8833daefffab84183ee9a4a38f5d3a30bb