Support wrap content height in RecyclerBinder
Summary:
If `mIsWrapContent` is enabled, we have to make sure that we have calculate the child height in `computeRangeLayout`, which is called when items are inserted / updated / removed.

After we have the child height in the range, we can trigger `measure()` again and use the child height to calculate the wrapped height.

Reviewed By: mihaelao

Differential Revision: D7499914

fbshipit-source-id: 131584ba7504fbad064ef1b56c0658d56a6199d5