CardSpec corner shadow and edge shadow are discontinuous
Summary:
CardSpec corner shadow and edge shadow are discontinuous

Top Shadow
{F123773246}

Bottom Shadow
{F123773253}

When we draw the corner, we were using the radius of corner to draw the corner of shadow which was wrong, the center of shadow gradient won't align with the base line of edge gradient. So it's end up with discontinuous between edge and corner

we should use (shadow_horizontal + corner_radius) as the shadow corner radius to align the gradient with the edge. In this way, both top and bottom are the same paint which is a radial gradient with (shadow_horizontal + corner_radius) radius.

Reviewed By: IanChilds

Differential Revision: D7632248

fbshipit-source-id: df13262b5441ddeb82a2c06bdf615c807233680c