Ensure new ViewPager pages are mounted
Summary:
When we add listeners to ViewPagers for incremental mount, we post it to avoid a ConcurrentModificationException which can occur if we are adding a new LithoView during an incremental mount that was caused by scrolling the ViewPager. However, this leads to a bug in the following scenario:
1, Start at tab 0 of the ViewPager.
2, Click on tab 2. This hasn't yet been created so the view is laid out. The ViewPager is still at position 0 so the content of tab 2 is not yet mounted.
3, Tab 2 posts the addition of a listener to the ViewPager.
4, The ViewPager is moved to position 2.
5, The listener is added to the ViewPager. This is too late to have any effect so no content is shown.

Instead, we want to add the listener immediately unless it we are already iterating the listeners. I can't think of a better way to do this that catching the exception.

Reviewed By: muraziz

Differential Revision: D6682957

fbshipit-source-id: 85c141ebc6aa09b7065684fa0fa21fc68e515aae