Collect transition information during measure when there is chance that LithoView width/height animation is going to run
Summary:
There is one big limitation of current transitions in which the first frame of LithoView height/width animation is skipped if transition is provided by state update or diff of props/state is used when declaring animation.

A bit of background: Top-level component which corresponds to LithoView when mounted treats height/width animations a bit different than other components. The reason for that is that LithoView is the bridge between component world and view world and when its bounds change it needs to inform view hierarchy above, therefore special handling needed (basically we call requestlayout() in every frame of LithoView bounds animation).

We apply animated value of LithoView bounds animation on measure, however, mount time and state update transitions are collected only in mount phase. Component is mounted to LithoView when its performLayout is called at which point measure is already called and we miss that first frame.

The solution to this is to collect transitions on measure if there is chance that LithoView width/height animation is going to run.

Reviewed By: astreet

Differential Revision: D7675744

fbshipit-source-id: de819b7513133e31055c50026792e534daa26801