Enable LithoView bounds animation when componentTree has changed but new one is not yet mounted.
Summary: We had a condition that if `LithoView` receives new `ComponentTree`, then we don't apply bounds animation during measure. There are cases where `LithoView` is receiving `ComponentTree` first time (hence its previous `ComponentTree` will be null and `mHasNewComponentTree` will be `true`) or `LithoView` is being recycled and it had some other `ComponentTree` before, but new `ComponentTree` is not yet mounted. In those cases we still want to apply bounds animation to LithoView as it might be appearing first time to user.

Reviewed By: astreet

Differential Revision: D7687307

fbshipit-source-id: a0fb65e927119266450afae5605dcaa2ea39a789