Add functionality for @CommonProp
Summary: One of the drawbacks with common props is that if you want to use it for something inside one of your methods, you have to add another method with a similar name and hope that people decide to use that one when using your component. Instead, it's much better to give people access to the common props that people have set on the component. To do this we just allow you to do `CommonProp` in the same way you can do `Prop`. Right now this just works for delegate methods in mount specs and layout specs - we can extend it to other specs (probably not a big deal since they have much fewer common props) and other methods in the future as and when required.

Reviewed By: passy, astreet

Differential Revision: D7708841

fbshipit-source-id: a17b7df0283133677823751553414fa35c09e452