Add new perf logger for mount
Summary:
This introduces a new perf logging interface which closer maps to our internal
representation. Importantly, it avoids allocating an intermediate LogEvent data
structure that is then used to copy over its data to another similar data
structure.

Secondly, it allows for the use of typed data and removes, for instance, the
unnecessary serialization to a JSON array for doubles. Given that the mount
events can fire up to 120 times a second (once per frame both for items entering
and leaving the screen) this has the potential to be impactful on perf.

N.B. The odd-looking `.toArray(new Boxed[0])` is in fact the most performant way
to convert a list to an array (see
https://shipilev.net/blog/2016/arrays-wisdom-ancients/). We cannot do that on
the receiver side as different boxed arrays erase to the same signature.

In order to assess the impact, the two logging modes will coexist while we run
an experiment.

Reviewed By: IanChilds

Differential Revision: D7745482

fbshipit-source-id: bdb377d4fd2295729de461540850f75d45e9ba84