Revert D6968169 (Fix animating LithoView bounds on first mount)
Summary: D6968169 was added because sample app was behaving abnormally (ExpandableElement sample showing blank when entering). Now that issue doesn't occur anymore even after reverting that and we have better ways of handling that specific scenario we can get rid of that change which was also making hard to make changes to bounds animation.

Reviewed By: IanChilds

Differential Revision: D7773625

fbshipit-source-id: 5789caf12ea937fa94c21eb4312fc7c4db9034ba