Implement async update
Summary: Adds async updates. These are a bit special: they will immediately update inserted items, and will update items pending insertion. This does break the batching behavior a bit, but the idea is that it's much harder to batch the update because the user can actually interact with that item in the meantime, meaning generating the layout async is not very feasible. If we have complaints or find issues, we'll explore other options.

Reviewed By: pasqualeanatriello

Differential Revision: D7761770

fbshipit-source-id: 78ae5bd615785b3c0f7cefec801901539b2c52ee