add KeyUpEvent, make EditText emit KeyUpEvent's
Summary:
Introducing KeyUpEvent and making EditText components emit those.
Scenario: there are multiple EditText's on the screen. When user clicks backspace while focused on an empty EditText, we want focus to change to a previous EditText. Seems like the only way to achieve this behavior is to use KeyEvents. This commit makes it possible

Reviewed By: muraziz

Differential Revision: D7773286

fbshipit-source-id: e5a73bf57f9a1f49b791bdaffd73172a6ad3e7d5