Don't crash for performIncrementalMount() with null layout state
Summary: In some cases we will indiscrimately call performIncrementalMount on all LithoViews within a ViewGroup (an example of this is when things move onto screen via some kind of animation that we don't get callbacks in LithoView for, so we just call performIncrementalMount() on all LithoViews that could have been affected. In these cases, if they don't have a main thread layout state then we can just bail early, we shouldn't crash.

Reviewed By: muraziz

Differential Revision: D7775317

fbshipit-source-id: f344d7214ecccd2ce2cad5c6b8c7344219897d53