Don't propagate info from parent ComponentContext in RecyclerBinder
Summary: this context is used to create the ComponentTrees inside RecyclerBinder. If we use the parent context, we'll propagate down info like the component scope, the key handler and state handler. The parent context belongs to a different tree though and these shouldn't be shared with the ComponentTrees in the binder

Reviewed By: IanChilds

Differential Revision: D7789327

fbshipit-source-id: 55e87049d176f21ed1f6291bc076f13269b02d4b