Fix leak coming from SnapHelper
Summary:
D6126185 added the ability to specify a SnapHelper directly, but this leads to leaks as the configuration is stored statically, and it has a reference to the snap helper which may reference a recycler view and not be unreferenced.

The reason that we allowed specifying a SnapHelper directly rather than just adding another SnapMode was that we thought a lot of people might want to do this, however in the intervening seven months it doesn't look like anyone else has so let's just remove the extra complexity and go back to having a SnapMode.

Reviewed By: muraziz

Differential Revision: D7844372

fbshipit-source-id: 4d346cc430141acfb97756628ad966746fb9498a