Support WRAP_CONTENT in RecyclerBinder
Summary:
Use ViewportFiller to calculate WRAP_CONTENT. The major change in RecyclerBinder is

1. In `measure()`, use `fillListViewport()` to calculate the wrapped size, it's only allowed to be executed multiple times in main thread only when `mWrapContent` is enabled.

2. In `measure()`, the measured size is passed into `fillListViewport()` as a parameter, and it might be modified when we finish calculating the wrapped size, thus after back in `measure()` we can use it to update the `outSize`.

3. When `notifyChangeSetComplete()` is called, trigger `requestRemeasure()` with `mRequiresRemeasure.set(true)` to remeasure the size without using the cached one.

 ---
About Supporting item decoration, because ItemDecorations are applied to the canvas in `RecyclerView#onDraw(Canvas)`, I'm thinking about adding a listener in the onDraw() method, at that time we can get correct decoration size through `getTopDecorationHeight` / `getBottomDecorationHeight` by LayoutManager. Then we can dispatch a RemeasureEvent to RecyclerBinder to remeasure with the decoration size.

Reviewed By: astreet

Differential Revision: D7687935

fbshipit-source-id: 768c7ea2f91a5da3378ad4a9c26578dfcb73a5bd