Make sure contentDescriptions properly work on Recycler components
Summary: Setting a contentDescription on a RecyclerView doesn't actually allow the view to be focusable and have the description read.  Instead, it should simply append the contentDescription to the existing "in list" and "out of list" announcements that are made when a user moves focus into and out of the list.  This previously wasn't possible in Litho, as there was no way to get the contentDescription from within the RecyclerSpec as it was a CommonProp.  As of D7708841 this is now possible, and this diff fixes the issue so that RecyclerSpec functions identically to RecyclerView.

Reviewed By: IanChilds

Differential Revision: D7836241

fbshipit-source-id: 86012eb56752b1babe74dcaaf54dc184bf0b9b9a