Disable item view cache for RecyclerView used in RecyclerSpec
Summary:
RecyclerView's ViewCache is supposed to be an optimization which allows views to go offscreen completely and come back to viewport without rebinding views. However, the assumptions there doesn't work well with RecyclerBinder which assumes that whatever comes to screen should be bound with ComponentTree instances that it holds. In most cases this works fine but there are edge cases where the LithoView which is in ViewCache loses its ComponentTree (as it could be reassigned to another LithoView) and when that LithoView comes to screen RV doesn't rebind it therefore we end up with 0-height LithoViews.

First we faced this issue in Messenger ThreadView and here is the diff (D7600480) in which I disabled item view cache and after public testing we found no regression on perf metrics therefore shipped it. Now we got more reports facing similar issue and we are going to disable view cache in `RecyclerSpec` component which uses RV under the hood.

Reviewed By: IanChilds

Differential Revision: D7908419

fbshipit-source-id: 1c7b8ee9d2908283de0be4fdf4b04fa776d33f1c