Compare to previous view bounds in LithoView.performLayout()
Summary: It looks like we had a bug here - we were looking to check that the bounds had changed, but l, t, r, b are given relative to the parent whereas the visible rect used for mount is given relative to the view itself. Therefore we also need to save off the view bounds when we mounted to decide whether to mount at the point or not.

Reviewed By: astreet

Differential Revision: D7909058

fbshipit-source-id: dafb8a1bd5fccd41e78075400410a58faccf7e68