Clean up TreeProps initialization
Summary:
This removes two leftovers from the old days of TreeProps:

**Resetting the treeprops based on the parent context**

This was added in D4294348 when the context of local scoping still existed and
children were recursively created, which the consequence of the local context
being mutated. This is no longer relevant as we create a new context for each
child. Hence, we don't need to restore the original context.

**Resetting the scoped tree props**

D4413043 did this as a mitigation for a memory leak in January 2017. Again, this
was for a time where we recursively passed around the context and could (in
cases like Layout.create or Wrapper.create) create a new layout based on a
Context which may hold a TreeProp. This, again, is no longer possible as we have
separate instances.

**Why change this?**

It complicates things, but more importantly, it means that we clear the
TreeProps before we reach the MOUNT stage, which blocks some logging
improvements for which we would like to have access to them.

There's also the same logic block in SectionsTree, but I'll tackle that
separately as there's no test coverage for this.

Reviewed By: IanChilds

Differential Revision: D7952243

fbshipit-source-id: 726833d1a5ad06df6a0447b8ea7ccc4f377e4e83