Add experiment to use hasOverlappingRendering=false
Summary:
The default hasOverlappingRendering returns true, which means that we will use an offscreen buffer to apply alpha rendering. This is very expensive for time spent on the GPU.

This test experiments with setting it to false so that the ComponentHost's children are directly drawn one after another with the corresponding alpha which is much cheaper. See https://www.facebook.com/notes/andy-street/react-native-android-offscreen-alpha-compositing/10153949352079590/ for more info on this.

Reviewed By: passy

Differential Revision: D7988588

fbshipit-source-id: a575a1f0923dbed7251c9d221c3db6d58a8dc212